+++ writeLines.js
@@ -1,6 +1,13 @@
 const fs = require('fs');
 const fsp = require('fs').promises;
 const path = require('path');
+const getLinesCount = require('./getLinesCount');
+
+const directoryExistsSync = require('./directoryExistsSync');
+const fileExistsSync = require('./fileExistsSync');
+
+const directoryExists = require('./directoryExists');
+const fileExists = require('./fileExists');
 
 /**
  * Writes a line or several lines to the end of the file.
@@ -11,64 +18,80 @@ const path = require('path');
 module.exports = async function writeLines(
   directoryPath,
   fileName,
-  arrayOfLinesofData
+  data
 ) {
-  let directoryExists = fs.existsSync(directoryPath);
-  let fileExists =
-    !fileName || !directoryExists
-      ? false
-      : await fs.existsSync(path.join(directoryPath, fileName));
 
-  if (!directoryExists) {
+  if (!directoryPath || !fileName){
+    return {
+      success: false,
+      message: 'Writelines expects both, a valid directory path and a file name.'
+    }
+  }
+
+  if (!data ) {
+    return {
+      success: false,
+      message: 'No data provided.'
+    }
+  }
+
+  let arrayOfLinesofData=[];
+  if (!Array.isArray(data)){
+    arrayOfLinesofData.push(data)
+  } else {
+    arrayOfLinesofData = [...data]
+  }
+
+  let directory_exists = await directoryExists(directoryPath);
+  let file_exists = await fileExists(directoryPath,fileName);
+
+  if (!directory_exists) {
     return {
       success: false,
       message: `Error: ${directoryPath} doesnot exist.`
     };
-  } else if (directoryExists && !fileExists) {
+  } 
+  
+  if (directory_exists && !file_exists) {
     return {
       success: false,
       message: `Error: ${path.join(directoryPath, fileName)} doesnot exist.`
     };
-  } else if (directoryExists && fileExists) {
-    let filehandle;
-    try {
-      // actual writing of line
-      const fullFilePath = path.join(directoryPath, fileName);
-      filehandle = await fsp.open(fullFilePath, 'a');
+  } 
+  
 
-      if (
-        arrayOfLinesofData &&
-        Array.isArray(arrayOfLinesofData) &&
-        arrayOfLinesofData.length > 0
-      ) {
-        arrayOfLinesofData.forEach(async data => {
-          await fsp.appendFile(filehandle, `\r\n${parsedData(data)}`);
-        });
-        return {
-          success: true,
-          message: `SUCCESS: ${arrayOfLinesofData.length} lines successfully written to ${fileName}.`
-        };
-      } else {
-        if (arrayOfLinesofData) {
-          await fsp.appendFile(
-            filehandle,
-            `\r\n${parsedData(arrayOfLinesofData)}`
-          );
-          return {
-            success: true,
-            message: `SUCCESS: 1 line successfully written to ${fileName}.`
-          };
+    // let filehandle;
+    try {
+      const fullFilePath = path.join(directoryPath, fileName);
+      // filehandle = await fsp.open(fullFilePath, 'a');
+
+      let linesCount = await getLinesCount(directoryPath, fileName);
+      arrayOfLinesofData.forEach( async (data, index) => {
+        console.log('data ', data);
+
+        if (linesCount && linesCount.data > 0){
+          await fsp.appendFile(fullFilePath, `\r\n${parsedData(data)}`);
+        } else {
+          if(index === 0){
+            await fsp.appendFile(fullFilePath, `${parsedData(data)}`);
+          } else {
+            await fsp.appendFile(fullFilePath, `\r\n${parsedData(data)}`);
+          }
         }
-      }
+      });
+      return {
+        success: true,
+        message: `SUCCESS: ${arrayOfLinesofData.length} lines successfully written to ${fileName}.`
+      };
     } catch (error) {
       return {
         success: false,
         message: error.toString()
       };
     } finally {
-      if (filehandle !== undefined) await filehandle.close();
+      // if (filehandle !== undefined) await filehandle.close();
     }
-  }
+
 };
 
 function parsedData(fileData) {